# PERF: cythonizing _concat_date_cols; conversion to float without exceptions in _does_string_look_like_datetime