@anmyachev 2019-03-17
I'm trying to speed up read_csv function. Manually, I ran asv as follows:
`asv continuous -f 1.05 origin HEAD -b io.csv`
And got these result:
       before           after         ratio
     [707c7201]       [846f0bfa]
     <master>         <datehelpers-performance-patch>
-     2.98±0.06ms      2.78±0.02ms     0.93  io.csv.ReadCSVFloatPrecision.time_read_csv_python_engine(',', '.', 'round_trip')
-     2.56±0.04ms      2.38±0.01ms     0.93  io.csv.ReadCSVFloatPrecision.time_read_csv_python_engine(';', '_', 'round_trip')
-     1.94±0.03ms      1.79±0.04ms     0.93  io.csv.ReadCSVDInferDatetimeFormat.time_read_csv(True, 'ymd')
-     3.27±0.08ms      3.02±0.04ms     0.92  io.csv.ReadUint64Integers.time_read_uint64
-     1.55±0.03ms      1.38±0.02ms     0.89  io.csv.ReadCSVDInferDatetimeFormat.time_read_csv(False, 'ymd')

During further testing performance, one of the datasets showed a decrease in the read_csv runtime for 25%.