@jreback 2019-03-17
these are tiny numbers - and likely not the cause of an actual slowdown - read_csv is io bound for the most part