@jreback 2019-03-18
@anmyachev can u show the effect on the asv benchmarks for csv