@anmyachev 2019-03-18
Original `csv` benchmarks were mostly not measuring `_concat_date_time` - most their time was spent in parsing item. I have composed new benchmarks (see 61d8e6e) that highlight the speedup brought by this PR.

master | datehelpers-performance-patch | speedup | test name
-|-|-|-
28.6±0.3ms |      24.2±0.2ms |    0.85 | io.csv.ReadCSVConcatDatetime.time_read_csv
12.1±0.2ms |      7.17±0.3ms |    0.59 | io.csv.ReadCSVConcatDatetimeBadDateValue.time_read_csv('0')
17.1±0.1ms |      6.90±0.1ms |    0.40 | io.csv.ReadCSVConcatDatetimeBadDateValue.time_read_csv('nan')
16.2±0.4ms |      5.04±0.2ms |    0.31 | io.csv.ReadCSVConcatDatetimeBadDateValue.time_read_csv('')