@vnlitvin 2019-03-18
> I would rather you cythonize things if need by.

I have implemented a small benchmark for `_does_string_look_like_datetime`. It clearly shows that Cython in certain cases like tight loops or lots of exception raising/handling can lag behind C code a bit:

before|           after|         ratio|test case
-|-|-|-
[707c7201]|       [e66883ef]||
**pre-patch**|       **datehelpers-performance-patch**||
105±3ms|      32.9±0.05ms|     0.31|  io.parsers.DoesStringLookLikeDatetime.time_check_datetimes('0.0')
168±4ms|         46.1±1ms|     0.27|  io.parsers.DoesStringLookLikeDatetime.time_check_datetimes('10000')
296±5ms|       41.5±0.4ms|     0.14|  io.parsers.DoesStringLookLikeDatetime.time_check_datetimes('2Q2005')