@vnlitvin 2019-03-18
Also note that benchmarks are currently working on quite limited amount of data to be fast, if I patch those to resemble real-life scenarios (using 5M lines instead of 50K) results would be:

before           |after         |ratio|test case
-|-|-|-
[707c7201]|       [e66883ef]||
**pre-patch**|       **datehelpers-performance-patch**||
3.38±0s|          2.77±0s|     0.82|  io.csv.ReadCSVConcatDatetime.time_read_csv
1.19±0.01s|          584±9ms|     0.49|  io.csv.ReadCSVConcatDatetimeBadDateValue.time_read_csv('0')
1.73±0.01s|         482±10ms|     0.28|  io.csv.ReadCSVConcatDatetimeBadDateValue.time_read_csv('nan')
1.66±0.01s|          414±2ms|     0.25|  io.csv.ReadCSVConcatDatetimeBadDateValue.time_read_csv('')

Which yields at least 0.5s saving on reading (or more than 1 second in the best case for our optimization).