@mroeschke 2019-03-18
Is there any way the to align `_does_string_look_like_datetime` closer to what your implementation in C does or are the speedups purely language specifics? I am not the biggest fan of the maintenance burden of this approach either.