@vnlitvin 2019-03-19
@mroeschke we did our best to bring our improvements from C to Cython for `_does_string_look_like_datetime` (and new benchmarks show it's now on par with C version).

We've now trying to re-write `_concat_date_cols` in Cython and we'll see how it would impact the benchmark numbers.