@pep8speaks 2019-03-22
Hello @anmyachev! Thanks for updating this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:





* In the file [`pandas/io/parsers.py`](https://github.com/pandas-dev/pandas/blob/2e2d58881364a7a078bfd21d1426ba0dd235cbe5/pandas/io/parsers.py):

> [Line 17:1](https://github.com/pandas-dev/pandas/blob/2e2d58881364a7a078bfd21d1426ba0dd235cbe5/pandas/io/parsers.py#L17): [E265](https://duckduckgo.com/?q=pep8%20E265) block comment should start with '# '