@vnlitvin 2019-04-08
With #25946 merged, here are fresh figures.
`asv continuous -f 1.01 upstream/master datehelpers-performance-patch -b io.csv -e -a sample_time=2 -a warmup_time=2` yields:

before|after|ratio|test name
-|-|-|-
[6de81336]|[1631aca2]||
**master**|**datehelpers-performance-patch**||
1.78±0.01ms|1.71±0.02ms|0.96|io.csv.ReadCSVDInferDatetimeFormat.time_read_csv(True, 'ymd')
1.74±0.02ms|1.66±0.01ms|0.95|io.csv.ReadCSVDInferDatetimeFormat.time_read_csv(True, 'iso8601')
1.40±0.01ms|1.28±0.01ms|0.91|io.csv.ReadCSVDInferDatetimeFormat.time_read_csv(False, 'ymd')
1.49±0.03ms|1.35±0ms|0.91|io.csv.ReadCSVDInferDatetimeFormat.time_read_csv(False, 'iso8601')
23.8±0.6ms|19.9±0.08ms|0.83|io.csv.ReadCSVConcatDatetime.time_read_csv
10.8±0.3ms|6.11±0.2ms|0.57|io.csv.ReadCSVConcatDatetimeBadDateValue.time_read_csv('0')
17.1±0.1ms|5.86±0.02ms|0.34|io.csv.ReadCSVConcatDatetimeBadDateValue.time_read_csv('nan')
15.0±0.1ms|4.17±0.2ms|0.28|io.csv.ReadCSVConcatDatetimeBadDateValue.time_read_csv('')